change type of iteration variable

+++ src/main.c
@@ -24,8 +24,8 @@ void example_1(void) {
 		fvec_push(v, (float) rand() / (float) RAND_MAX);
 	}
 	printf("nitems: %ld\n", fvec_count(v));
-	for (int i = 0; i < 5; i++) {
-		printf("%f\n", fvec_get(v, (size_t) i));
+	for (size_t i = 0; i < 5; i++) {
+		printf("%f\n", fvec_get(v, i));
 	}
 	fvec_free(&v);
 }
@@ -42,8 +42,8 @@ void example_2(void) {
 	uvec_push(v, (User){ 125, "Bob" });
 	uvec_push(v, (User){ 300, "Alice" });
 
-	for (int i = 0; i < uvec_count(v); ++i)
-		printf("User's name: %s\n", uvec_get(v, (size_t) i).name);
+	for (size_t i = 0; i < uvec_count(v); ++i)
+		printf("User's name: %s\n", uvec_get(v, i).name);
 
 	uvec_free(&v);
 }
@@ -54,8 +54,8 @@ void example_3(void) {
 	int ns[] = { 32, 17, 5, 286, 114, 0, 2, 7 };
 	ivec_t v = ivec_from_array(ns, ARR_LEN(ns));
 
-	for (int i = 0; i < ivec_count(v); ++i)
-		printf("%d\n", ivec_get(v, (size_t) i));
+	for (size_t i = 0; i < ivec_count(v); ++i)
+		printf("%d\n", ivec_get(v, i));
 
 
 	ivec_free(&v);
